test-requirements.txt
pytest>=3,<4
pytest-flake8==1.0.4
pytest-mock==0.10.1
pytest-isort==0.2.1
pytest-black==0.3.5
pytest-mypy==0.3.2
pytest-aiohttp==0.3.0
pytest-asyncio==0.10.0
pytest-random-order==1.0.4
mypy==0.670
black==19.3b0
